+++ packages/teamscale-coverage-collector/src/main.ts
@@ -11,7 +11,6 @@ import axios from 'axios';
 import FormData from 'form-data';
 import QueryParameters from './utils/QueryParameters';
 import { inspect } from 'util';
-import tmp from 'tmp';
 import mkdirp from 'mkdirp';
 import path from 'path';
 import { StdConsoleLogger } from './utils/StdConsoleLogger';
@@ -25,9 +24,11 @@ import { PrettyFileLogger } from './utils/PrettyFileLogger';
  */
 type Parameters = {
 	// eslint-disable-next-line camelcase
-	dump_to_file?: string;
+	dump_to_folder: string;
 	// eslint-disable-next-line camelcase
 	log_to_file: string;
+	// eslint-disable-next-line camelcase
+	keep_coverage_files: boolean;
 	// eslint-disable-next-line camelcase
 	log_level: string;
 	// eslint-disable-next-line camelcase
@@ -56,10 +57,20 @@ type Parameters = {
 };
 
 /**
+ * Error that is thrown when the upload to Teamscale failed
+ */
+class TeamscaleUploadError extends Error {
+	// No special fields needed compared to Error.
+	// Sole use is to be able to distinguish Teamscale upload errors from other errors.
+}
+
+/**
  * The main class of the Teamscale JavaScript Collector.
  * Used to start the collector for with a given configuration.
  */
 export class Main {
+	private static readonly DEFAULT_COVERAGE_LOCATION = 'coverage';
+
 	/**
 	 * Construct the object for parsing the command line arguments.
 	 */
@@ -72,7 +83,15 @@ export class Main {
 
 		parser.add_argument('-v', '--version', { action: 'version', version });
 		parser.add_argument('-p', '--port', { help: 'The port to receive coverage information on.', default: 54678 });
-		parser.add_argument('-f', '--dump-to-file', { help: 'Target file to write coverage to.' });
+		parser.add_argument('-f', '--dump-to-folder', {
+			help: 'Target folder for coverage files.',
+			default: this.DEFAULT_COVERAGE_LOCATION
+		});
+		parser.add_argument('-k', '--keep-coverage-files', {
+			help: 'Whether to keep the coverage files on disk after a successful upload to Teamsacle',
+			action: 'store_true',
+			default: false
+		});
 		parser.add_argument('-l', '--log-to-file', { help: 'Log file', default: 'logs/collector-combined.log' });
 		parser.add_argument('-e', '--log-level', { help: 'Log level', default: 'info' });
 		parser.add_argument('-t', '--dump-after-mins', {
@@ -154,12 +173,12 @@ export class Main {
 				{ level: logLevel, stream: new PrettyFileLogger(fs.createWriteStream(logfilePath)), type: 'raw' }
 			]
 		});
-    
+
 		// If the given flag is set, we also log with a JSON-like format
 		if (config.json_log) {
 			logger.addStream({ level: logLevel, path: `${logfilePath}.json` });
 		}
-    
+
 		return logger;
 	}
 
@@ -174,12 +193,6 @@ export class Main {
 		const logger = this.buildLogger(config);
 		logger.info(`Starting collector in working directory "${process.cwd()}".`);
 		logger.info(`Logging "${config.log_level}" to "${config.log_to_file}".`);
-
-		// Check the command line arguments
-		if (!config.dump_to_file && !config.teamscale_server_url) {
-			logger.error('The Collector must be configured to either dump to a file or upload to Teamscale.');
-			process.exit(1);
-		}
 
 		// Prepare the storage and the server
 		const storage = new DataStorage(logger);
@@ -227,31 +240,34 @@ export class Main {
 
 	private static async dumpCoverage(config: Parameters, storage: DataStorage, logger: Logger): Promise<void> {
 		try {
-			const deleteCoverageFileAfterUpload = !config.dump_to_file;
-			const coverageFile = config.dump_to_file ?? tmp.tmpNameSync();
-			try {
-				// 1. Write coverage to a file
-				const lines = storage.dumpToSimpleCoverageFile(coverageFile);
-				logger.info(`Dumped ${lines} lines of coverage to ${coverageFile}.`);
+			// 1. Write coverage to a file
+			const [coverageFile, lines] = storage.dumpToSimpleCoverageFile(config.dump_to_folder, new Date());
+			logger.info(`Dumped ${lines} lines of coverage to ${coverageFile}.`);
 
-				// 2. Upload to Teamscale if configured
-				if (config.teamscale_server_url) {
-					await this.uploadToTeamscale(config, logger, coverageFile, lines);
-				}
-			} finally {
-				if (deleteCoverageFileAfterUpload) {
+			// 2. Upload to Teamscale if configured
+			if (config.teamscale_server_url) {
+				await this.uploadToTeamscale(config, logger, coverageFile, lines);
+				// Delete coverage if upload was successful and keeping coverage files on disk was not configure by the user
+				if (!config.keep_coverage_files) {
 					fs.unlinkSync(coverageFile);
 				}
 			}
 		} catch (e) {
-			logger.error('Coverage dump failed.', e);
+			if (e instanceof TeamscaleUploadError) {
+				logger.error(
+					`Teamscale upload failed. The coverage files on disk (inside the folder "${config.dump_to_folder}") were not deleted. 
+					You can still upload them manually.`,
+					e
+				);
+			} else {
+				logger.error('Coverage dump failed.', e);
+			}
 		}
 	}
 
 	private static async uploadToTeamscale(config: Parameters, logger: Logger, coverageFile: string, lines: number) {
 		if (!(config.teamscale_access_token && config.teamscale_user && config.teamscale_server_url)) {
-			logger.error('Cannot upload to Teamscale: API key and user name must be configured!');
-			return;
+			throw new TeamscaleUploadError('API key and user name must be configured!');
 		}
 
 		if (lines === 0) {
@@ -260,27 +276,27 @@ export class Main {
 
 		logger.info('Preparing upload to Teamscale');
 
-		const form = new FormData();
-		form.append('report', fs.createReadStream(coverageFile), 'coverage.simple');
+		const form = this.prepareFormData(coverageFile);
+		const queryParameters = this.prepareQueryParameters(config);
+		await this.performTeamscaleUpload(config, queryParameters, form, logger);
+	}
 
-		const parameters = new QueryParameters();
-		parameters.addIfDefined('format', 'SIMPLE');
-		parameters.addIfDefined('message', config.teamscale_message);
-		parameters.addIfDefined('repository', config.teamscale_repository);
-		parameters.addIfDefined('t', config.teamscale_commit);
-		parameters.addIfDefined('revision', config.teamscale_revision);
-		parameters.addIfDefined('partition', config.teamscale_partition);
-
+	private static async performTeamscaleUpload(
+		config: Parameters,
+		parameters: QueryParameters,
+		form: FormData,
+		logger: Logger
+	) {
 		await axios
 			.post(
-				`${config.teamscale_server_url.replace(/\/$/, '')}/api/projects/${
+				`${config.teamscale_server_url?.replace(/\/$/, '')}/api/projects/${
 					config.teamscale_project
 				}/external-analysis/session/auto-create/report?${parameters.toString()}`,
 				form,
 				{
 					auth: {
-						username: config.teamscale_user,
-						password: config.teamscale_access_token
+						username: config.teamscale_user ?? 'no username provided',
+						password: config.teamscale_access_token ?? 'no password provided'
 					},
 					headers: {
 						Accept: '*/*',
@@ -292,22 +308,44 @@ export class Main {
 				if (error.response) {
 					const response = error.response;
 					if (response.status >= 400) {
-						logger.error(`Upload failed with code ${response.status}: ${response.statusText}`);
-						logger.error(`Request failed with following response: ${response.data}`);
+						throw new TeamscaleUploadError(
+							`Upload failed with code ${response.status}: ${response.statusText}. Response Data: ${response.data}`
+						);
 					} else {
 						logger.info(`Upload with status code ${response.status} finished.`);
 					}
 				} else if (error.request) {
-					logger.error(`Upload request did not receive a response.`);
+					throw new TeamscaleUploadError(`Upload request did not receive a response.`);
 				}
 
 				if (error.message) {
-					logger.error(`Something went wrong when uploading data: ${error.message}`);
-					logger.debug(`Details of the error: ${inspect(error)}`);
+					logger.debug(
+						`Something went wrong when uploading data: ${error.message}. Details of the error: ${inspect(
+							error
+						)}`
+					);
+					throw new TeamscaleUploadError(`Something went wrong when uploading data: ${error.message}`);
 				} else {
-					logger.error(`Something went wrong when uploading data: ${inspect(error)}`);
+					throw new TeamscaleUploadError(`Something went wrong when uploading data: ${inspect(error)}`);
 				}
 			});
+	}
+
+	private static prepareQueryParameters(config: Parameters) {
+		const parameters = new QueryParameters();
+		parameters.addIfDefined('format', 'SIMPLE');
+		parameters.addIfDefined('message', config.teamscale_message);
+		parameters.addIfDefined('repository', config.teamscale_repository);
+		parameters.addIfDefined('t', config.teamscale_commit);
+		parameters.addIfDefined('revision', config.teamscale_revision);
+		parameters.addIfDefined('partition', config.teamscale_partition);
+		return parameters;
+	}
+
+	private static prepareFormData(coverageFile: string) {
+		const form = new FormData();
+		form.append('report', fs.createReadStream(coverageFile), 'coverage.simple');
+		return form;
 	}
 }