+++ test/systemtest.mjs
@@ -1,10 +1,9 @@
 import LocalWebServer from 'local-web-server';
 import cypress from 'cypress';
-import { execSync } from 'child_process';
+import { execSync, spawn } from 'child_process';
 import path from 'path';
 import * as fs from 'fs';
 import * as fsp from 'fs/promises';
-import { spawn } from 'child_process';
 import ServerMock from 'mock-http-server';
 
 /**
@@ -170,10 +169,10 @@ function identifyExpectedButAbsent(actual, expected) {
  * Start the collector process.
  * @returns {ChildProcessWithoutNullStreams}
  */
-function startCollector(coverageTargetFile, logTargetFile, projectId) {
+function startCollector(coverageFolder, logTargetFile, projectId) {
 	const collector = spawn(
 		'node',
-		[`${COLLECTOR_DIR}/dist/src/main.js`, `-f`, `${coverageTargetFile}`, `-l`, `${logTargetFile}`, `-e`, `info`],
+		[`${COLLECTOR_DIR}/dist/src/main.js`, `-f`, `${coverageFolder}`, `-l`, `${logTargetFile}`, `-k`, `-e`, `info`],
 		{
 			env: {
 				...process.env,
@@ -259,13 +258,14 @@ for (const study of caseStudies) {
 		});
 
 		try {
-			const coverageTargetFile = path.resolve(`${study.rootDir}/coverage.simple`);
+			const coverageFolder = path.resolve(path.join(study.rootDir, 'coverage'));
 			const logTargetFile = path.resolve(`${study.rootDir}/coverage.log`);
 
 			// Delete existing files
-			if (fs.existsSync(coverageTargetFile)) {
-				await fsp.unlink(coverageTargetFile);
+			if (fs.existsSync(coverageFolder)) {
+				fs.rmdirSync(coverageFolder, { recursive: true });
 			}
+			fs.mkdirSync(coverageFolder);
 			if (fs.existsSync(logTargetFile)) {
 				await fsp.unlink(logTargetFile);
 			}
@@ -282,7 +282,7 @@ for (const study of caseStudies) {
 				}
 			});
 
-			await new Promise((resolve, reject) => {
+			await new Promise(resolve => {
 				console.log('## Starting the Teamscale mock server');
 				teamscaleServerMock.start(resolve);
 			});
@@ -290,7 +290,7 @@ for (const study of caseStudies) {
 			// Start the new collector
 			console.log('## Starting the collector');
 			console.log(`Collector is logging to ${logTargetFile}`);
-			const collectProcess = startCollector(coverageTargetFile, logTargetFile, study.name);
+			const collectProcess = startCollector(coverageFolder, logTargetFile, study.name);
 
 			try {
 				console.log('## Running Cypress');
@@ -319,13 +319,18 @@ for (const study of caseStudies) {
 			if (!fs.existsSync(logTargetFile)) {
 				throw new Error('The coverage collector is supposed to write a log file!');
 			}
-			if (!fs.existsSync(coverageTargetFile)) {
+
+			const coverageFiles = fs.readdirSync(coverageFolder);
+			if (coverageFiles.length === 0) {
 				throw new Error('The coverage collector did not write a coverage file!');
+			} else if (coverageFiles.length > 1) {
+				throw new Error('More than one coverage file was written');
 			}
+			const coverageFile = path.join(coverageFolder, coverageFiles[0]);
 
 			// Analyze the coverage
-			console.log(`## Analysis of coverage in ${coverageTargetFile}`);
-			checkCoverage(coverageTargetFile, study);
+			console.log(`## Analysis of coverage in ${coverageFile}`);
+			checkCoverage(coverageFile, study);
 
 			// Check the calls to the Teamscale mock server
 			const mockedRequests = teamscaleServerMock.requests({ method: 'POST' }).length;
@@ -335,7 +340,7 @@ for (const study of caseStudies) {
 				console.log(`Received ${mockedRequests} requests in the Teamscale mock server.`);
 			}
 
-			await new Promise((resolve, reject) => {
+			await new Promise(resolve => {
 				console.log('## Stopping the Teamscale mock server');
 				teamscaleServerMock.stop(resolve);
 			});